pyproject.toml
[project]
name = "m-haihu2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "numpy<2",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
]
